Constructing a URL Dynamically | The Reverse Function | Named URLs
Description:

If for instance you have hard coded a url like /challenges in views.py and you change the /challenges to /challengessss in urls.py, then this will cause a break until you change  /challenges to /challengessss in views.py as well. So to avoid this we use named url with reverse function to dynamically construct a url.