Adding a Slug Field | Overwriting save() Method | Updating Migrations
Description:

We added a new model field called slug, this will populate the slug filed for new data, but what about existing data? For that we have to update slug for existing data manually. So to do so, for example we want to update object with id=4,  you can just run <Book.objects.get(id=4).save()> on python shell, save will trigger on this object automatically, and slug field will be there for this existing object too.