Relations | onetomany | Creating Author Model | Dealing with Migrations Stuff
Description:

In this lesson we are trying to make author field as a foreign key, as we are going to define Author aa standalone entity. But while trying this we'll face certain DB problems. Like what will happen to those rows which already has author set as a string names? Because certainly now author will hold an integer value as a foreign key.

So the best answer is plan your DB design ahead of the time. So that in latter point of time we don't face these problems. For now we are going to delete < Book.objects.all().delete() > all the records in book table, and start with empty table.